requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2
